address comments. refactor FilterBar > SavedFilter
