+++ src/pages/view/FilterBar/FilterBar.js
@@ -1,7 +1,7 @@
 import { useDhis2ConnectionStatus } from '@dhis2/app-runtime'
 import i18n from '@dhis2/d2-i18n'
 import PropTypes from 'prop-types'
-import React, { useCallback, useEffect, useMemo, useState } from 'react'
+import React, { useState } from 'react'
 import { connect } from 'react-redux'
 import {
     acRemoveItemFilter,
@@ -11,41 +11,17 @@ import ConfirmActionDialog from '../../../components/ConfirmActionDialog.js'
 import { sGetNamedItemFilters } from '../../../reducers/itemFilters.js'
 import FilterBadge from './FilterBadge.js'
 import classes from './styles/FilterBar.module.css'
-import { Button, colors, FlyoutMenu, IconMore16 } from '@dhis2/ui'
-import {
-    acSetActiveFilter,
-    tDeleteActiveFilter,
-    tSaveFilter,
-    tToggleActiveFilterVisibility,
-} from '../../../actions/savedFilters'
-import { useCachedDataQuery } from '@dhis2/analytics'
-import {
-    privateVisiblity,
-    sGetActiveFilter,
-} from '../../../reducers/savedFilters'
-import { deepEqual } from '../../../modules/util'
-import MenuItem from '../../../components/MenuItemWithTooltip'
-import DropdownButton from '../../../components/DropdownButton/DropdownButton'
-import SaveFilterDialog from './SaveFilterDialog'
-import { isFilterActionAllowed } from '../../../api/savedFilters'
+import { sGetActiveFilter } from '../../../reducers/savedFilters'
+import SavedFiltersBlock from './SavedFilterBlock'
 
 const FilterBar = ({
     filters,
+    activeFilter,
     removeFilter,
     removeAllFilters,
-    saveFilter,
-    updateActiveFilter,
-    deleteFilter,
-    activeFilter,
-    toggleFilterVisibility,
 }) => {
     const { isConnected: online } = useDhis2ConnectionStatus()
-    const { currentUser } = useCachedDataQuery()
     const [dialogIsOpen, setDialogIsOpen] = useState(false)
-    const [filterDialogIsOpen, setFilterDialogIsOpen] = useState(false)
-    const [filterDialogData, setFilterDialogData] = useState(activeFilter)
-    const [moreOptionsIsOpen, setMoreOptionsIsOpen] = useState(false)
-    const [showScope, setShowScope] = useState(true)
 
     const onRemoveFilter = (filterId) => {
         if (!online && filters.length > 1) {
@@ -57,120 +33,10 @@ const FilterBar = ({
 
     const closeDialog = () => setDialogIsOpen(false)
 
-    const closeFilterDialog = useCallback(() => {
-        setFilterDialogIsOpen(false)
-        setShowScope(true)
-    }, [activeFilter])
-    const onConfirmFilterDialog = useCallback(
-        async ({ name, visibility, id }) => {
-            await handleSaveFilter({ ...activeFilter, name, visibility, id })
-            closeFilterDialog()
-        },
-        [activeFilter, closeFilterDialog]
-    )
-
-    const hasActiveSavedFilter = Boolean(activeFilter.id)
-    const savedFilterHasChanges = useMemo(
-        () => hasActiveSavedFilter && !deepEqual(activeFilter.values, filters),
-        [activeFilter, filters]
-    )
-
-    const toggleMoreActions = () => {
-        setMoreOptionsIsOpen((prev) => !prev)
-    }
-    const showFilterAction = isFilterActionAllowed(activeFilter, currentUser)
-
-    const handleSaveFilter = useCallback(async (filter = {}) => {
-        await saveFilter(currentUser, {...activeFilter, ...filter})
-        setMoreOptionsIsOpen(false)
-    }, [currentUser, activeFilter])
-    const handleSaveNewFilter = useCallback(() => {
-        setFilterDialogData({ id: 'new' })
-        setFilterDialogIsOpen(true)
-    }, [activeFilter])
-    const handleRenameFilter = useCallback(() => {
-        setShowScope(false)
-        setFilterDialogData(activeFilter)
-        setFilterDialogIsOpen(true)
-        setMoreOptionsIsOpen(false)
-    }, [activeFilter])
-    const handleDeleteFilter = useCallback(async () => {
-        await deleteFilter(currentUser)
-        setMoreOptionsIsOpen(false)
-    }, [currentUser])
-    const handleToggleVisibility = useCallback(async () => {
-        await toggleFilterVisibility(currentUser)
-        setMoreOptionsIsOpen(false)
-    }, [currentUser])
-
-    useEffect(() => {
-        if (!filters.length) {
-            updateActiveFilter(null)
-        }
-    }, [filters])
-
-    const getMoreActions = useMemo(
-        () => (
-            <FlyoutMenu>
-                {showFilterAction && (
-                    <>
-                        <MenuItem
-                            dense
-                            label={i18n.t('Rename')}
-                            onClick={handleRenameFilter}
-                        />
-                        <MenuItem
-                            dense
-                            label={i18n.t('Delete')}
-                            onClick={handleDeleteFilter}
-                        />
-                    </>
-                )}
-                {savedFilterHasChanges && (
-                    <>
-                        {showFilterAction &&
-                            <MenuItem
-                                dense
-                                label={i18n.t('Save')}
-                                onClick={handleSaveFilter}
-                            />
-                        }
-                        <MenuItem
-                            dense
-                            label={i18n.t('Save as new filter')}
-                            onClick={handleSaveNewFilter}
-                        />
-                    </>
-                )}
-                {showFilterAction &&
-                    <MenuItem
-                        dense
-                        label={
-                            activeFilter?.visibility === privateVisiblity
-                                ? i18n.t('Make it public')
-                                : i18n.t('Make it private')
-                        }
-                        onClick={handleToggleVisibility}
-                    />
-                }
-            </FlyoutMenu>
-        ),
-        [
-            savedFilterHasChanges,
-            handleRenameFilter,
-            handleDeleteFilter,
-            handleSaveNewFilter,
-            handleSaveFilter,
-            handleToggleVisibility,
-            activeFilter,
-            showFilterAction,
-        ]
-    )
-
     return filters.length ? (
         <>
             <div className={classes.bar} style={{ alignItems: 'center' }}>
-                {hasActiveSavedFilter && <div>{activeFilter.name}:</div>}
+                {activeFilter.id && <div>{activeFilter.name}:</div>}
                 {filters.map((filter) => (
                     <FilterBadge
                         key={filter.id}
@@ -178,26 +44,7 @@ const FilterBar = ({
                         onRemove={onRemoveFilter}
                     />
                 ))}
-                {online && !hasActiveSavedFilter && (
-                    <Button secondary small onClick={handleSaveNewFilter}>
-                        {i18n.t('Save')}
-                    </Button>
-                )}
-                {online && hasActiveSavedFilter && (showFilterAction || savedFilterHasChanges) && (
-                    <DropdownButton
-                        dataTest="more-actions-button"
-                        secondary
-                        small
-                        showArrow={false}
-                        open={moreOptionsIsOpen}
-                        disabledWhenOffline={true}
-                        onClick={toggleMoreActions}
-                        icon={<IconMore16 color={colors.grey700} />}
-                        component={getMoreActions}
-                    >
-                        <wbr />
-                    </DropdownButton>
-                )}
+                <SavedFiltersBlock />
             </div>
             <ConfirmActionDialog
                 open={dialogIsOpen}
@@ -210,27 +57,15 @@ const FilterBar = ({
                 onConfirm={removeAllFilters}
                 onCancel={closeDialog}
             />
-            {filterDialogIsOpen && (
-                <SaveFilterDialog
-                    open={filterDialogIsOpen}
-                    filter={filterDialogData}
-                    showScope={showScope}
-                    onCancel={closeFilterDialog}
-                    onConfirm={onConfirmFilterDialog}
-                />
-            )}
         </>
     ) : null
 }
 
 FilterBar.propTypes = {
     filters: PropTypes.array.isRequired,
+    activeFilter: PropTypes.object,
     removeAllFilters: PropTypes.func.isRequired,
     removeFilter: PropTypes.func.isRequired,
-    updateActiveFilter: PropTypes.func.isRequired,
-    deleteFilter: PropTypes.func.isRequired,
-    toggleFilterVisibility: PropTypes.func.isRequired,
-    activeFilter: PropTypes.object,
 }
 
 FilterBar.defaultProps = {
@@ -245,8 +80,4 @@ const mapStateToProps = (state) => ({
 export default connect(mapStateToProps, {
     removeAllFilters: acClearItemFilters,
     removeFilter: acRemoveItemFilter,
-    saveFilter: tSaveFilter,
-    updateActiveFilter: acSetActiveFilter,
-    deleteFilter: tDeleteActiveFilter,
-    toggleFilterVisibility: tToggleActiveFilterVisibility,
 })(FilterBar)
