+++ src/reducers/index.js
@@ -14,7 +14,7 @@ import selected from './selected.js'
 import showDescription from './showDescription.js'
 import slideshow from './slideshow.js'
 import visualizations from './visualizations.js'
-import savedFilters from "./savedFilters";
+import savedFilters from "./savedFilters.js";
 
 export default combineReducers({
     dashboards,
